add a EULA that user's can read and accept
Accepting the EULA should be made optinal once the OpenAI components
are also made optional. That way people can play without needing to
use it. If we do end up working on a local model, the EULA can be
repalced with a policy noting that no information is collected, and
thus can be come a notice instead of a requirement.